Customize submission mail

+++ db/migrate/20170223170808_delete_all_mail_validated_in_db.rb
@@ -1,5 +1,6 @@
 class DeleteAllMailValidatedInDb < ActiveRecord::Migration[5.0]
   def change
-    MailTemplate.where(type: "MailValidated").delete_all
+    mail_template_exist = Object.const_get(:MailTemplate).is_a?(Class) rescue false
+    MailTemplate.where(type: "MailValidated").delete_all if mail_template_exist
   end
 end
